+++ src/renderer/render/media/image.ts
@@ -9,9 +9,12 @@ export function image(target: HTMLImageElement, url: URL, cache?: Collection<str
       .map(attr => [attr.name, attr.value])));
   define(target, {
     src: target.getAttribute('data-src'),
-    style: 'max-width: 100%;',
     loading: 'lazy',
   });
-  cache?.set(url.href, target.cloneNode(true));
+  cache?.set(url.href, define(target.cloneNode(true), {
+    width: null,
+    height: null,
+    'aspect-ratio': null,
+  }));
   return target;
 }